@@ -1,8 +1,8 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from "@angular/core";
+import { BrowserModule } from "@angular/platform-browser";
 
-import { AppComponent } from './app.component';
-import {OnBoardComputerModule} from "./on-board-computer/on-board-computer.module";
+import { AppComponent } from "./app.component";
+import { OnBoardComputerModule } from "./on-board-computer/on-board-computer.module";
 
 @NgModule({
   declarations: [
@@ -15,4 +15,5 @@ import {OnBoardComputerModule} from "./on-board-computer/on-board-computer.modul
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+}
