@@ -1,19 +1,24 @@
-import { Module } from '@nestjs/common';
-import { ControllersModule } from 'controllers/controllers.module';
-import { ConfigModule } from '@nestjs/config';
-import universeJson from './config/configuration';
-import { ServicesModule } from './services/services.module';
+import { MiddlewareConsumer, Module, NestModule } from "@nestjs/common";
+import { ControllersModule } from "controllers/controllers.module";
+import { ConfigModule } from "@nestjs/config";
+import universeJson from "config/configuration";
+import { ServicesModule } from "services/services.module";
+import { MissionOddMiddleware } from "middlewares/mission-odd.middleware";
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       load: [universeJson],
-      isGlobal: true,
+      isGlobal: true
     }),
     ServicesModule,
-    ControllersModule,
+    ControllersModule
   ],
   controllers: [],
-  providers: [],
+  providers: []
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(MissionOddMiddleware).forRoutes("v1/mission");
+  }
+}
