@@ -1,8 +1,8 @@
-import { CommandFactory } from 'nest-commander';
-import { CliModule } from 'cli/cli.module';
+import { CommandFactory } from "nest-commander";
+import { CliModule } from "cli/cli.module";
 
 async function bootstrap() {
-  await CommandFactory.run(CliModule, ['warn', 'error']);
+  await CommandFactory.run(CliModule, ["warn", "error"]);
 }
 
 bootstrap();
