@@ -1,1 +1,3 @@
-export const ODD_PROPERLY_COMPUTE = 'The mission odd is properly computed';
+export const ODD_PROPERLY_COMPUTE = "The odd mission is properly computed";
+export const IMPORTED_FILE_ISSUE = "Invalid provided file";
+export const INTERNAL_ERROR = "Issue while computing the odd";
