@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
 import { ServicesModule } from "services/services.module";
-import { MissionOddController } from "controllers/mission-odd/mission-odd.controller";
+import { OddMissionController } from "controllers/odd-mission/odd-mission.controller";
 import { OddMissionService } from "services/odd-mission/odd-mission.service";
 
 @Module({
   imports: [ServicesModule],
   providers: [OddMissionService],
-  controllers: [MissionOddController]
+  controllers: [OddMissionController]
 })
 export class ControllersModule {
 }
