@@ -1,12 +1,12 @@
+import { Planet } from "models/planet.model";
+
 /**
  * The Galaxy
  * - planets: Listing the planets of the galaxy
  *
  * @author Laurent
- * @version 1
+ * @version 1.0
  */
-import { Planet } from 'models/planet.model';
-
 export class Galaxy {
   private readonly _planets: Planet[];
 
@@ -15,8 +15,9 @@ export class Galaxy {
   }
 
   /**
-   * Does the galaxy contains planet
+   * Does the galaxy contains a given planet
    * @param planetName {string} Planet name
+   * @return {boolean} state the existence
    */
   doesContainsPlanet(planetName: string): boolean {
     return this._planets
