@@ -1,4 +1,4 @@
-import { Module } from "@nestjs/common";
+import { CacheModule, Module } from "@nestjs/common";
 import { EmpireService } from "services/empire/empire.service";
 import { OddMissionService } from "services/odd-mission/odd-mission.service";
 import { RepositoryModule } from "repositories/repositoryModule";
@@ -6,7 +6,10 @@ import { RoutesService } from "services/routes/routes.service";
 import { UniverseService } from "services/universe/universe.service";
 
 @Module({
-  imports: [RepositoryModule],
+  imports: [RepositoryModule, CacheModule.register({
+    max: 2,
+    ttl: 600000 // 10 min
+  })],
   providers: [
     EmpireService,
     RoutesService,
